Minor type changes

+++ dlib/dnn/cpu_dlib.cpp
@@ -1639,11 +1639,11 @@ namespace dlib
             float* g = grad.host();
             const float x_scale = (grad.nc()-1)/(float)std::max<long>((gradient_input.nc()-1),1);
             const float y_scale = (grad.nr()-1)/(float)std::max<long>((gradient_input.nr()-1),1);
-            for (long samp = 0; samp < gradient_input.num_samples(); ++samp)
+            for (long long samp = 0; samp < gradient_input.num_samples(); ++samp)
             {
-                for (long k = 0; k < gradient_input.k(); ++k)
+                for (long long k = 0; k < gradient_input.k(); ++k)
                 {
-                    for (long r = 0; r < gradient_input.nr(); ++r)
+                    for (long long r = 0; r < gradient_input.nr(); ++r)
                     {
                         const float y = r*y_scale;
                         const long long top    = static_cast<long long>(std::floor(y));
